move NewReturned to goc from goprint
